refactor: extract folder template services
